On Linux, handle all CAN captures with pcap-linux.c, in cooked mode.
There's no need to support capturing on PF_CAN sockets; CAN interfaces
show up as regular interfaces on which you can capture with PF_PACKET
sockets, so just let pcap-linux.c handle them, and get rid of
pcap-can-linux.c.

Capture on them in cooked mode, so we get the protocol field and can
distinguish between "classic" CAN and CAN FD.

The hardware for which pcap-canusb-linux.c was intended never reached
production:

    https://github.com/axos88/libpcap/commit/f3edbb599b8cbcc7e4560000dcba8e992dc11a31#commitcomment-18716617

so we don't need pcap-canusb-linux.c, either.

This all removes the need for the "host-endian" link-layer header type;
it never made it into a libpcap release, so we just remove it.  The
"big-endian" link-layer header type is kept for the benefit of packets
captured with the old pcap-can-linux.c code; we revert it to its old
name.