Fix indentation, clean up strcmp() calls.
If you want to compare strings with a given C comparison operator, you
are expected to do strcmp(str1, str2) {operator} 0; that makes it
clearer what comparison is being done.  "!x" checks for x being zero,
but it's mainly intended for testing Boolean variables; using it on
strcmp() makes it a bit figure out, as comparing against 0 means
comparing for "equal", not for "not equal".