Clean up findalldevs code.
Have a routine that unconditionally adds a device to the list of
devices, without bothering to check whether there's already a device
with that name, and a separate routine that does the check and, if it
doesn't find the device, calls the routine to add ti.  That avoids
scanning the entire list in cases where we know the search will fail.
The only reasons for doing the check are that we're on a platform where
we find the list of interfaces by a call that returns a list of
*addresses*, with an interface name attached to each address, and

	1) for each address, we need to make sure we don't already have
	   the interface, create it if we don't, and add the address to
	   the now-guaranteed-to-exist entry for the interface;

	2) we might have to make a *separate* call to enumerate
	   interfaces, to find interfaces with *no* addresses, and that
	   call also enumerates the ones that *do* have addresses, and
	   we don't want to create a duplicate entry for them.

Change some findalldevs helper routines to have "dev" rather than "if"
in the name, as not all devices are regular network interfaces.

For the DAG findalldevs routine, make sure it always provides an error
message if it fails.

We don't need add_addr_to_iflist() or get_if_description() on Windows,
so don't define them on Windows.

Update comments to reflect reality.