Better handle errors opening USB monitor devices.
If we get EACCES, return PCAP_ERROR_PERM_DENIED, to indicate that we
don't have permission to do the capture.

If we get an error other than ENOENT on the binary device, fail
immediately, as it means there *is* support for doing a binary capture,
it's just that some problem occurred when trying it.

In either of those cases, don't fall back on the text device.