Clean up the binary-vs-text stuff.
Base the decision as to whether to use binary or text mode on the kernel
version - if we're running on 2.6.21 or later, use binary mode and don't
bother with text mode, as we have binary mode if we have USB monitoring
at all, and if we're running on a pre-2.6.21 kernel, use text mode and
don't bother with text mode, as we definitely don't have binary mode.

That way, on 2.6.21, we won't list any usbmon devices if we don't have
USB monitoring (which probably means "the usbmon module isn't loaded"),
which hopefully makes it a bit clearer what the problem is.  We'll still
list those devices prior to 2.6.21, but there doesn't seem to be a good
way to test for the presence of USB monitoring support that works if
monitoring support is built into the kernel (in which case it's present)
or if it's a module (in which case it's present only if the module is
loaded) and that doesn't require root privileges (we don't want the
enumeration of devices to require root privileges, as then "you don't
have sufficient privilege and the program doing the work doesn't have it
either" presents as "why am I not seeing devices?" rather than "it tells
me I don't have sufficient permission").