Have separate "find device" and "find or add device" routines.
Use the "find device" routine for CSPi's Myricom adapters; those
adapters can appear as regular network interfaces, but, when we
enumerate the Myricom adapters, we want to update the description for
those adapters, rather than adding separate entries for them.  While
we're at it, add comments to packet-snf.c explaining this.

Fix some comments and some indentation while we're at it.