Round the buffer size up to a multiple of the frame size.
When calculating the number of frames for the memory-mapped buffer,
round the buffer size up to a multiple of the frame size - rather than
rounding down, which is what happens by default with a divide, as that
could give a buffer smaller than our caller asked for, all the way down
to "zero frames" if the buffer size is less than the frame size.

Addresses the main problem with GitHub issue #548.