Call the timeout the "packet buffer timeout" rather than the "read timeout".
That should make it clearer that the timeout does *not* guarantee that a
call that reads packets will return within N milliseconds even if no
packets arrive.

In the pcap_open_live() and pcap_set_timeout() calls, point to the
detailed description of the packet buffer timeout in pcap(3PCAP).