Some fixes for the new API.
Save the result of the add_dev() call, so that the "#if 0" code could be
made to work.  (If it can't be made to work, it should be removed, along
with the curdev variable and the assignment to it.)

Don't pass dev->flags to add_dev(), as that routine expects PCAP_IF_
flags, not IFF_ flags - and we don't have the relevant IFF_ flags anyway
(IFF_UP, IFF_RUNNING, IFF_LOOPBACK if that applies).