In pcap_compile(), first check whether the pcap_t is activated.
Before we allocate or otherwise set up anything, check whether the
pcap_t is activated, and set the error message string and return -1 if
it's not.

That way, we don't go through the cleanup code in that code path -
there's nothing to clean up.

Fixes the issue in GitHub pull request #552.