Support setting non-blocking mode before activating.
We just set a flag and attempt to set non-blocking mode after
activating.

If a module can't support non-blocking mode, it should set the set
non-blocking operator in the create routine, so a pre-activation call
will fail the same way a post-activation call fails.

While we're at it:

Have the get non-blocking and set non-blocking modes not take an error
buffer as an argument; they have the error buffer in the pcap_t to set.
pcap_getnonblock() and pcap_setnonblock() just copy the error from there
to the argument passed in.

Make sure we set the cleanup op pointer when appropriate.