Fix some build warnings.
ctrlport is an array, so it needn't be checked against NULL.

In UN*X, the last argument to getsockopt() is a pointer to socklen_t; in
Windows, it's a pointer to int.  On Windows, sockutils.h typedefs
socklen_t as int, so use socklen_t as the type of the variable.

In C, for historical reasons, a function with no arguments must be
declared or defined with "void" as the argument list - if there's no
argument list, it's an old-style no-prototype function.