Check for zero-length addresses.
At least on my macOS Sierra machine, the utun0 device has a destination
address with an sa_len of 0 (and an sa_family of AF_UNSPEC).  That's not
valid - sa_len is *supposed* to be the total length of the structure,
but a total length of 0 doesn't even leave room for the sa_len and
sa_family fields, which are present in *all* sockaddr structures.