Don't treat non-AF_INET addresses as AF_INET6 addresses.
An address that's not AF_INET is not necessarily AF_INET6.

While we're at it, clean up the code that processes the various
addresses in an address entry - let rpcap_deseraddr() do the checks for
AF_INET/AF_INET6 and, if neither the address nor the netmask nor the
broadcast address nor the destination address were allocated, discard
the entry, otherwise add it to the lsit of entries.