Fix pcap_lookupnet() on Windows.
Set the address table size variable to the full size of the table, so
that we get more than one address if there is more than one address.

Use the resulting address table size variable's value as the loop bound
when scanning over the addresses, so we don't process addresses that
aren't there.