Use compiler #defines to determine how to handle some function attributes.
Take the set of tests in the test programs that check compiler and
version to decide how to mark functions as printflike or as not
returning, and put them into a header.  Use that throughout the libpcap
code, rather than doing configuration-time tests that have to be
replicated in the autoconf script and CMake files.

That obviates the need to put those tests into CMakeLists.txt and means
that, if we ever make functions with those attributes available by code
outside libpcap (for example, if we ever introduce a plugin mechanism
for libpcap, to allow third-party pcap modules to be maintained
separately from libpcap), we can use those declarations (making a public
header file dependent on the results of a configure script usually
doesn't work well, for a variety of reasons).