bound snaplen for linux tpacket_v2 to ~64k
The aggregated ethernet packets produced by the linux kernel
can't exceed 65535 (plus eth overhead), so we can bound
accordingly the snaplen for tpacket_v2 on such devices.

This allows a more efficient usage of the receive buffer;

tcpdump has increased the default snaplen up to 256K in recent
versions, to cope better with USB data, and that leads to
tpacket_v2 rx ring being configured by default with tp_frame_nr
equal to 7, which is very bad for most use-case.

With this change the default tpacket_v2 layout has tp_frame_nr
equal to 32, which is still quite low, but works much better
and is consistent with the previous tcpdump behavior.

v1 -> v2:
	update the comment to better reflect the current status,
	code cleanup

Signed-off-by: Paolo Abeni <pabeni@redhat.com>