Don't reject pcap files or pcapng IDBs with a too-large snaplen.
Instead:

For pcap, don't allocate a buffer bigger than MAXIMUM_SNAPLEN, and let
the existing buffer size checks handle rejecting packets bigger than
MAXIMUM_SNAPLEN.

For pcapng, don't allocate a buffer bigger than a buffer big enough to
have a MAXIMUM_SNAPLEN-sized packet and 128KB of options, never grow the
buffer bigger than that, and fail if we see a packet that won't fit in a
maximum-sized buffer or with a captured length that's larger than the
snapshot length.

This should address GitHub issue #527.