Make VLAN filter handle both metadata and inline tags
Commit 04660eb1e561 ("Use BPF extensions in compiled filters") taught
libpcap to use VLAN BPF extensions to implement "vlan" and "vlan <id>"
filters. While this fixed filtering in the case when VLAN information
is in packet metadata (which is more frequent on Linux with recent
kernels), it broke the case when outermost VLAN tag is still in packet
data. Unfortunately, even if kernel supports BPF VLAN extensions, we have
to expect both types of packets.

This means that if the initial SK_AD_VLAN_TAG_PRESENT test is negative, we
need to also check for inline VLAN tag. As this test is done only for some
packets (determined at run time), we need to update variable part of link
type and link payload offsets, not constant part.