Add pcap_set_protocol() to allow specifying a specific capture protocol.
This adds the Linux-specific pcap_set_protocol(), which allows you to
specify a specific protocol to be captured.  A protocol value specified
with this function will be used instead of ETH_P_ALL when creating and
binding a packet socket.

For ethernet packets, a similar result can be achieved by explicitly
creating a BPF filter that filters on ethertype and attaching that
filter to the packet socket, however, filtering using the packet socket
protocol field is slightly more efficient.

Also, in some cases, the packet socket protocol field and the
ethertype field have different values, and in those cases it is
impossible to achieve the same result by using a BPF filter.  An
example of this is on Arista platforms, where sFlow-sampled packets
are delivered, unmodified, to virtual interfaces on the management
CPU with their packet socket protocol value set to the magic value
0x002d, independently of their actual ethertype.