Clean up some warnings.
Throw in some casts for comparisons, *and* make sure the values being
cast are within the range of the target type.

This means checking for a buffer returned by getmsg() having a length of
-1, which, according to the HP-UX getmsg() man page, means there isn't
any data for that buffer.  Treat that as an error.

If something can have an unsigned int value, make it an unsigned int.