Fix test for message running past the end of the buffer.
len is no longer a variable used in the loop, so it doesn't reflect how
much data is left in the buffer after all the messages that have been
processed; compare the message length against ep - bp instead.

Remove a test that will always fail - at that point in the loop, we know
that bp < ep, so we know that ep - bp is > 0.

Add a check for pcap_breakloop() in the loop that reads buffers of
messages, as is done in other modules.

Put back the check to make sure that, before processing each message, we
have at least one minimum-sized message's worth of data left in the
buffer.

Add a bunch of comments.

Make the return statements use the no-parenthesis style used elsewhere
in this file.