More details.
Explicitly note that it's used when creating a capture socket.

Note that the values are typically values from <linux/if_ether.h>, and
that they are in host byte order.

Note that it does not work on devices other than network interfaces; for
example, it has no effect on Linux USB, Bluetooth, or netfilter devices.