Always check for EOF after reading a character from /etc/ethers.
We can get an EOF at any time, either due to a file that doesn't end
cleanly with an NL or due to an I/O error.  In pcap_next_etherent(),
every time we call a routine that returns a character from the file,
check for EOF and return NULL.

(According to the C90 standard (and presumably all subsequent standards),
1) you can safely pass EOF to all the isXXX() macros and 2) EOF is not a
space character, so we don't need to check for EOF there.)

This should address the concerns of Coverity CID 1416964; hopefully
Coverity realizes that getc() returns either a nonnegative value or EOF,
and therefore that checking for EOF is sufficient to ensure that the
return value is nonnegative.  (Given how much it complains about
strcmp() with a string constant, because some tests can be evaluated at
compile time and therefore some code happens to be unreachable in that
particular call - code that will be optimized out by any competent
compiler - I'm not going to *assume* Coverity has a clue there.)