Move URL parsing and construction to pcap.c.
Eventually, there will be new APIs allowing opening of remote interfaces
in a create/activate style; they will do URL parsing and will be in
pcap.c.  This is part of that, as well as part of a shorter-term project
to replace pcap_parsesrcstr()'s sscanf-based URL parsing with something
not using sscanf(), because we want to avoid MSVC complaints about
sscanf's insecurity without just replacing sscanf with sscanf_s(), as
their APIs are not the same.