Make it possible to build on Windows without packet.dll.
Rename pcap-win32.c to pcap-npf.c, in case other capture mechanisms are
available on Windows in the future.

If we don't find packet.dll, build with pcap-null.c, to make a libpcap
that can read and write capture files but can't capture.

Use the generic pcap_lookupdev() if we don't have packet.dll and aren't
building on MS-DOS.