Don't test for capture devices, test for header files.
Testing for /dev/bpf or /dev/bpfN in CMake requires that they be
readable.  Test for various header files required by various pcap-XXX.c
files, and use that to determine what the capture mechanism for the
host platform is.

Don't look for headers by absolute path, as that could get in the way of
cross-development (if any cross-development where the build machine and
the host machine are running different OSes works).

This means we have to determine whether we have a cloning BPF device at
run time; do so.