Use strtok_r as strtok_s under MinGW as well
With cmake, strtok_r isn't tested for on windows and assumed not to exist.
Newer version of MinGW have implemented it. Older version (e. g. the one provided by Debian) however seem to only have it under strtok_s.
It's probably a ok to stick with strtok_s for all of windows for now.