Don't wire the library name into pcap-config.
With autotools, set the package name with AC_INIT(), which now takes the
package name and version as arguments, rather than the name of a file in
the source directory.  That sets PACKAGE_NAME.

With CMake, set PACKAGE_NAME to ${LIBRARY_NAME}.

In pcap-config.in, use "-l@PACKAGE_NAME@", rather than -lpcap, as the
library name.

That means that pcap-config will use -lwpcap, rather than -lpcap, on
Windows, where the library's name is wpcap rather than just pcap.