crypt() isn't necessarily in libcrypt.
It's not in libcrypt in macOS, and may not be in libcrypt in other
UN*Xes.  First look for it in the system libraries, and then, only if
that fails, look for it in libcrypt.  If *that* fails, warn that we
won't be able to build rpcapd.

This means that the macOS builds with remote capturing enabled should
work; revert to requiring that they succeed.