Base the shadow-password decision on whether we have getspnam().
It appears to have the same signature and documented behavior in Linux
and Solaris, so I suspect that, if it exists on a given UN*X, it works
the same way.

We still need to handle password verification on other platforms that
don't store the encrypted password in the publicly-readable(!)
/etc/passwd or another publicly-readable form (i.e., probably just about
every UN*X on the planet, these days) but don't offer getspnam().