Plug some memory leaks.
We allocate memory for the user name and password when using
username/password authentication; free it if the authentication succeeds
or fails.

Fixes Coverity CIDs 1419022 and 1419023.

While we're at it, rename variables to indicate what they're doing
(they're not just two random strings, they're a user name and password),
and put the string pointers into the scope of the "username/password
authentication" case, just as the string lengths are.