Clean up reading error string from RPCAP_MSG_ERROR messages.
Add *another* needed va_end().

When receiving the error string, don't use the same buffer for the error
string and any error message returned by sock_recv(); it might be safe
to do so, but it's not very clean to assume that.

If the error string length is 0, don't read anything, just provide an
empty string.