socket() and accept() return INVALID_SOCKET on errors.
On Windows, you're supposed to check against INVALID_SOCKET.  On UN*X,
you're supposed to check against -1, but, on UN*X, we define
INVALID_SOCKET to be -1 so you can compare against INVALID_SOCKET on
both platforms.

That means that sock_open() should return INVALID_SOCKET on errors as
well.