Fixed wrong quote usage in rc files
The usage of quotes rc files is not very straight forward.
PACKAGE_VERSION_DLL may not be put in quotes except when used for FileVersion. There, it must be in quotes or windres will complain.
PACKAGE_NAME and PACKAGE_VERSION are already carring quotes when arriving form config.h.
So no need to put them in quotes again.