pcap-rdmasniff: Return correct error when ibv_get_cq_event() fails
When ibv_get_cq_event() fails, that means that a read() on the underlying fd
failed, which usually indicates that a signal was received.  If errno is EINTR
then we should check if handle->break_loop has been set.  If set, we return
PCAP_ERROR_BREAK, otherwise just retry.  This handles both SIGINT due to a
user stopping tcpdump with ctrl-C and also SIGALRM for periodic stats reporting.

Other errors "shouldn't happen" so we return PCAP_ERROR immediately, since our
state is probably messed up in a way we can't recover from.