Use pthread_equal() to compare thread IDs
Some pthread implementations define pthread_t as a structure rather than a pointer or an integer.

Perfectly fine according to The Open Group Base Specifications.
What is not fine is for daemon_serviceloop() to assume it to always be an integer.

So to improve portability, pthread_equal() should be used to compare thread IDs.