Redo the message processing in the server.
Have two service loops, one of which runs before we are authenticated
(so the only messages we accept are "authenticate me" messages) and one
which runs after we are authenticated (where we can accept command
messages, but not "authenticate me" messages).

Have the service loops do protocol version checking; currently, we only
support one version, but the first loop would do protocol negotiation in
the future, while the second loop would treat messages with a version
other than the negotiated one as an error (and there might be different
loops for different versions, as some or all messages might have
different formats with different versions).

They already need a switch statement to switch on the message type, so
we don't need to have rpcap_checkmsg() do the type checking.

Have routines to process the various message types; they do all the work
of processing the message, and return 0 on success, send an error
message to the client and return 0 on failure, and log an error message
and return -1 on a network error.

Check for errors when receiving and sending network data; always log
those messages.

Check to make sure the message length isn't too short; send an error
message to the client, and discard the message, if it is.