Fix up the negotiation, eliminating a failure mode with old clients.
If we send a "wrong version number" error message to an old client
(before the changes to support version negotiation) with a version
number other than 0, it'll treat it as a protocol error and return the
wrong error message to the caller.  Instead, if the client's version
number is too old for the server, send back the version number they sent
us; they'll try it again and get the same error, and give up.  (This
would happen only if there's a server that doesn't support version 0;
let's not create so many protocol versions that there's a temptation to
drop older ones.)

Also, add details on protocol versioning and protocol negotiation (the
construction of which pointed out the failure mode).