Have a variable that keeps track of whether we have a thread or not.
pthread_t is an opaque type; it's not guaranteed to be an integer or a
pointer, so we're not guaranteed to be able to set it to 0 and have that
mean "we don't have a thread".  (Heck, even if it *is* an integer,
there's no guarantee that pthread_create() will never return 0.)

Instead, have a separate variable that keeps track of whether we have a
thread.