Make sure rpcap_msg_err() returns a null-terminated string.
Make sure we null-terminate the string before discarding the rest of a
too-long error message fails.

While we're at it, if we get a network error, we don't take any special
shutdown action for the connection, so no need to have rpcap_msg_err()
return a status; either it's reporting an error given to us by the
server or it's reporting a network error.