Clean up error handling when closing an active connection.
If we get an error, return -1, and just report the first error; if we
got an error sending RPCAP_MSG_CLOSE to the server, don't check for
errors when closing the connection.

For calls where we're not going to check for an error, don't pass an
error message buffer.

Addresses Coverity CID 1419000.