Clean up packet message reception.
Have separate sockutils routines for reading from a stream socket and
receiving messages from a datagram socket.  The latter assumes you get
the entire message in a single receive operation and checks for the
datagram being bigger than the provided buffer.

We don't need separate network and packet-to-provide-to-the-callback
buffers; use the standard buffer for both purposes, making it big enough
for the largest expected packet message.  This eliminates a copy.

Use the "receive messages from a datagram socket" routine for UDP data
sockets.  That way, we check for messages bigger than the largest
expected packet message; add additional checks to make sure the packet
is big enough to hold an rpcap header and, if so, that the packet length
in the rpcap header isn't larger than the part of the message.

Check for interrupted message receives in the "read a packet message"
code path.

If we get an invalid message from the server, discard the rest of it if
it's sent over TCP.

In the read routine, check for errors and, if we didn't get a packet and
didn't get an error - which could be a timeout or an interrupted receive
- check whether we were told to break out of the loop.