Clean up handing of signal interrupts in pcap_read_nocb_remote().
If the select() is interrupted by a signal, just return 0, don't treat
it as an error.  That'll let the read routine check whether we're
supposed to break out of the loop and, if not, it'll let it keep
processing packets, as we do for local capture modules.

When reading packet messages from a TCP socket, use the bp and cc
members of the pcap_t to keep track of where we're reading data and the
number of bytes of data we have read so far.  If a read is interrupted
by a signal, just return 0, as per the above.  Don't assume we're at the
beginning of a message when we enter pcap_read_nocb_remote(); instead,
if we've read less than the size of a packet message header, try to read
the rest of the header, otherwise extract the packet payload length from
the header, make sure it's not larger than the largest payload length
we're expecting (if it is, return an error), and try to read the rest of
the packet.  That way, we can handle multiple calls to
pcap_read_nocb_remote() reading multiple bits of the same packet, due to
reads being interrupted.

That obviates the need to discard the rest of the packet, as we read all
of it into the buffer.