Don't link with -lsocket and -lxnet on Solaris.
I don't know what happens if you use both, e.g. does recvmsg() support
the Single UNIX Specification stuff that the XPG 4.2 version from -lxnet
does but the default version from -lsocket doesn't?

That means we don't get getifaddrs(), but that appears to be based on
SIOCGLIFCONF on Solaris, and we already support using SIOCGLIFCONF, so
let's leave out a not-very-useful middleman.