In some older Solarises, link with even more libraries.
On Solaris, we have to link with -lxnet to get a version of recvmsg()
that will tell the caller if a datagram was truncated.  However, on some
older Solarises, we have to link with -lsocket and -lnsl to get
inet_ntop() and inet_pton(), so link those test programs that use those
routines with -lxnet *and* -lsocket and -lnsl if necessary.

(We don't worry about those older systems with CMake.)