Have a separate Makefile for the test programs.
Put that Makefile in the tests directory, and build the tests in that
directory.

This makes it easier to have different compiler flags for the library
and the test programs, and also makes the build procedure with autotools
a bit more like the build procedure for CMake.

Update the Travis tests to match, and do the findalldevstest test with
both autotools and with CMake.