Fix, and better explain, the "IPv6 means IPv6, not IPv4" option setting.
Use it if IPV6_V6ONLY (the RFC 3493 name) or IPV6_BINDV6ONLY are
defined, rather than expecting BSD to be defined on systems that support
this (Linux, at least in later versions, and NT 6 and later support it
as well).  Use IPV6_V6ONLY and, if IPV6_V6ONLY isn't defined but
IPV6_BINDV6ONLY, define IPV6_V6ONLY as IPV6_BINDV6ONLY.

Oh, and actually *set* the on-the-stack variable used to turn the option
on.

Explain why we're doing this, giving RFC references, and noting that
some stacks *don't* support the feature we're turning off, so we can't
just rely on it and not bother creating a separate IPv4 socket bound to
INADDR_ANY and the rpcap port.