Fix "make depend" on platforms without BPF or PF_PACKET sockets.
We can't even generate dependencies for valgrindtest on those platforms,
as it'll either fail with "Unknown platform or platform that doesn't
support Valgrind" or fail because there's no bpf.h (e.g., on Solaris 10,
where it's Solaris but doesn't have BPF).  Therefore, don't include it
in the source if we aren't going to build it.