Explicitly warn that negative packet buffer timeouts should not be used.
We won't treat them as errors, because it appears:

	https://stackoverflow.com/questions/46569127/what-does-to-ms-1-do-on-winpcap-and-linux-libpcap/47689659#47689659

that some programs set it to -1, probably expecting it to put the pcap_t
in non-blocking mode (as a result of the *internal implementation* of
capture on a regular network interface on *some* platforms).  However,
it won't work on all platforms; if you want non-blocking mode, use
pcap_setnonblock().