Free the socket temporary when appropriate.
If we're using threads, free it if the attempt to create the thread
fails.

If we're using subprocesses, free it in the parent process, which isn't
using it.

(We don't really teed the socket temporary if we're using subprocesses,
as the socket value won't be overwritten - the multiple socket opens are
done in separate processes.)

This should fix Coverity CID 1418994.