Try to compile *and link* programs testing for reentrant APIs.
That way, even if the header doesn't declare the function, so there's no
prototype checking, the test will fail if the function isn't in the
standard libraries.  That fixes builds on HP-UX.

Also, use (const char *)0, rather than NULL, as the first argument to
those API calls in the tests, because NULL isn't necessarily defined.