We don't need to check for getaddrinfo twice.
AC_LBL_LIBRARY_NET now checks for it and fails if it's not found, as we
require it.  The check for IPv6 support can thus simply assume it's
present, and thus that it can enable IPv6 support unless the user
explicitly asks not to provide it.  (Is that option still worth
providing?)

Update a comment while we're at it.