Don't use fopen_s().
It always opens the file with "deny all" sharing on Windows; we don't
want that, because we want UN*X semantics, with "deny none" sharing -
that's what people expect.

We'll squelch the warnings later; fopen() does the same parameter
checking that fopen_s() does, it doesn't return an errno value but it
probably sets errno and errno is probably thread-safe (inferred from
Microsoft's documentation), and we don't *want* the "security" offered
by opening the file "deny all", so the only thing we don't want is the
warnings.